requirements.txt
aiogram==3.17.0
python-dotenv==1.0.0
aiohttp==3.11.11
pydantic==2.10.6
typing-extensions>=4.12.2